Added handling of the Runtime error for nofit case
For some of the peaks detected in first one show bad results in second unit.
So to ensure smooth running of the code, when such a error pops up, the exception is
handled by returning the guess values and a zero cov matrix.

Also the start and end peak handling was improved.