Added the code to select peaks and confirm shots
A new program to filter the peaks and confirm shot is added.
The program compares the parameter with their boundary value to see of the value fitted is reasonable
and it also compares the parameter across units to since shot si expected to be similar in different instruments

The plotting routine is modified to accept the mask created by shot_selector and filter the peaks.