SHOT DETECTION AND ENERGY BAND RATIOS: CZTI
These functionswere used to try to detect shots in CZTI energy band 20-40 keV
and to plot the ratios between shot peaks in 40-80/20-40keV and 80-100/20-40keV.
The ratios were plotted for both : shots detected in LAXPC (by time matching)
as well as an independent detection run on CZTI 20-40 keV light curve
for 3 sigma and 5 sigma peaks. There are no significant results as of this moment
for any of these.