Updated the codes to allow running the codes together
Updated the codes to incorporate the arg parser which allows the codes to accept the parameters
from the terminal line itself. Now the bash script can be invoked to run the codes in a batch with
codes accepting the output from the previous codes. The arguments to the code are kept same to ensure
the coherence.

Further commit may use variable names so that the new runs are smoother.

Also, the code now only fits for 3 parameters.

 Changes to be committed:
	modified:   my_funcs.py
	modified:   scripted_shot_algorithm.sh
	modified:   shot_detector.py
	modified:   shot_fitting.py
	modified:   shot_selector.py
	modified:   shot_splitter.py