Edited the fitter function to return the bound vals
In the earlier version, the fitter would return the guess value in case of no-fit.
Since the guess values lies in the middle of the range, selector code will not be able to detect such cases.

So to accommodate that, bound values will be returned in case of no fit.